Move Wallet definition to /models folder
